split ActiveCall in others components
